slub: avoid duplicate creation on the first object
When a kmem_cache is created with ctor, each object in the kmem_cache will
be initialized before ready to use.  While in slub implementation, the
first object will be initialized twice.

This patch reduces the duplication of initialization of the first object.

Fix commit 7656c72b ("SLUB: add macros for scanning objects in a slab").

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Acked-by: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>