UPSTREAM: regulator: core: Get voltage from parent if not available
Load switches are modeled as regulators but they just provide
the voltage of their parent input supply. So the drivers for
these switches usually don't provide a .get_voltage function
handler but there is code in the kernel that assumes that all
regulators should be able to provide its current voltage rail.

So, if the output voltage for a regulator is not available and
it has a parent supply, then pass the voltage of its parent.

BUG=chrome-os-partner:30167
TEST=IO Voltage domain patches can now get WiFi voltage.

Change-Id: Ib4668d51257f3d65b45d5c3f3ad9d3cc9a966775
Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>
Signed-off-by: Mark Brown <broonie@linaro.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit e303996e94b8705c85f3d78f3c094d05b0620c9d)
Reviewed-on: https://chromium-review.googlesource.com/214594
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>