prfm
Fix for prfm error on AArch64

The following error is generated with the mainline clang compiler
for AArch64:

arch/arm64/include/asm/processor.h:149:15:
error: invalid operand in inline asm: 'prfm pldl1keep, ${0:a}
'

Per comments by Tim Northover on the LLVM Bug database:

"It's rather unclear how it's better than "prfm pstl1keep, [%0]" though.
Not all instructions can make use of any offset, so wouldn't we have to be
conservative and always map it to "[xN]"?

This test works with GCC under unit test but not in the kernel build.

Not-Signed-off-by: Mark Charlebois <charlebm@gmail.com>