mm, oom: remove unnecessary exit_state check
The oom killer scans each process and determines whether it is eligible for oom
kill or whether the oom killer should abort because of concurrent memory
freeing.  It will abort when an eligible process is found to have TIF_MEMDIE
set, meaning it has already been oom killed and we're waiting for it to exit.

Processes with task->mm == NULL should not be considered because they are either
kthreads or have already detached their memory and killing them would not lead
to memory freeing.  That memory is only freed after exit_mm() has returned,
however, and not when task->mm is first set to NULL.

Clear TIF_MEMDIE after exit_mm()'s mmput() so that an oom killed process is no
longer considered for oom kill, but only until exit_mm() has returned.  This was
fragile in the past because it relied on exit_notify() to be reached before no
longer considering TIF_MEMDIE processes.

Signed-off-by: David Rientjes <rientjes@google.com>
Cc: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>