CHROMIUM: video: tegra: get limit voltage matching target gk20a rate
When increase gk20a freq rate, it will try to set voltage first,
when it calculated a target voltage, it will call
gk20a_dvfs_rail_apply_limits() to check if the voltage between
min/max_volt. But in this function, it use current freq to get
max_volt, not the target freq rate, so it will cause problems,
it mean the gk20a may run with lower voltage when increase freq.
This change use the target rate to scan the limited values to
get the right voltage.

BUG=chrome-os-partner:30499
TEST=added debug messages to check the voltage when running gpu
dvfs, the valuses are expected.

Change-Id: If6047f2cd1d5fa8f553757f14dd278faa5366668
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/210631
Tested-by: Wei Ni <wni.nvidia@gmail.com>
Tested-by: Kary Jin <karyj@nvidia.com>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>