FROMLIST: ARM: dts: Enable emmc and sdmmc on the rk3288-evb boards
This adds support for the sdmmc and emmc ports on the rk3288 using the
currently posted driver from Addy at:
  https://patchwork.kernel.org/patch/4653631/

This enables basic SD and eMMC support.  Things are not yet running at
the fastest speed and we don't have the regulators specified, but we
can at least use the eMMC and SD cards now.

A few notes:
* This is not baesd on Jaehoon's patch series removing the slot node,
  but it does use new syntax like putting the bus width at the top
  level and using the new cap-mmc-highspeed / cap-sd-highspeed.  A
  future patch will modify this one to remove the slot node.

BUG=chrome-os-partner:30167
TEST=With series, dw_mmc is working pretty well.

Change-Id: I64bab391dfa810beb26377be6a22d06b5d1fa4e7
Signed-off-by: Doug Anderson <dianders@chromium.org>
Acked-by: Arnd Bergmann <arnd@arndb.de>
(am from https://patchwork.kernel.org/patch/4657491/)
Reviewed-on: https://chromium-review.googlesource.com/208459