CHROMIUM: security: tlk_driver: add FIQ glue
Add FIQ glue for ARM64 cpus and register a handler with the secure world (TLK).
TLK handles the FIQ interrupts and calls our handler after it is done saving
the context. We then get the FIQ context via a SMC call and pass it to the
FIQ handler registered by the kernel (using fiq_glue_register_handler).

The FIQ debugger calls fiq_glue_register_handler() to register its handler
which we in turn call whenever a FIQ fires. The glue saves all the registers,
elr/spsr values and SP, which is then used by tlk_fiq_handler() to pass on to
the actual handler.

BUG=chrome-os-partner:42905
TEST=Tested SMC calls using TLK testapp

Change-Id: I6711c25892284dd82c9042479381eca5f707c9b4
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Signed-off-by: Mahesh Lagadapati <mlagadapati@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/287719
Commit-Ready: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>