CHROMIUM: drivers: thermal: tegra132-soctherm: add T132-specific SOC_THERM driver
Add a Tegra132-specific SOC_THERM driver that attempts to work around
the various thermal calibration issues with early versions of these
chips.

IMPORTANT NOTE: This driver is only meant to provide support for
THERMTRIP functionality at this point in time.  Thermal throttling
support is not yet implemented in the SOC_THERM drivers for T132, since
it differs from T124.

BUG=chrome-os-partner:30834
TEST=can read temperatures from thermal sysfs

Change-Id: If87be8be2562ef3204a75237710231809663ead2
Signed-off-by: Paul Walmsley <pwalmsley@nvidia.com>
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/212434
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>