FROMLIST: common: dma-mapping: Introduce common remapping functions
For architectures without coherent DMA, memory for DMA may
need to be remapped with coherent attributes. Factor out
the the remapping code from arm and put it in a
common location to reduce code duplication.

As part of this, the arm APIs are now migrated away from
ioremap_page_range to the common APIs which use map_vm_area for remapping.
This should be an equivalent change and using map_vm_area is more
correct as ioremap_page_range is intended to bring in io addresses
into the cpu space and not regular kernel managed memory.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>
(am from https://patchwork.kernel.org/patch/4710101
[djkurtz: map_vm_area(.., &pages) b/c chromeos-3.14 does not have:
 commit f6f8ed47353597dcb895eb4a15a28af657392e72
 mm/vmalloc.c: clean up map_vm_area third argument]
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:31704
TEST=compile tested on veyron & rush_ryu
TEST=boot tested on rush_ryu w/ v7 series
TEST=boot tested on veyron_pinky w/ v7 series + iommu/drm

Change-Id: I289a2256ebfb80de78522def50b4844923d0eb6d
Reviewed-on: https://chromium-review.googlesource.com/219500
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>