CHROMIUM: ARM: tegra: save/restore reset handler for suspend
The reset handler was stored in IRAM. And it will be powered off by
vdd_core in the LP0. This patch provide functions for LP0 to
save/restore reset handler.

At suspend time, the copy of the reset handler in DRAM should be
consistent with the one in IRAM, so there's no reason to copy it
from IRAM back to DRAM.  This was causing a page fault because the
reset handler in DRAM is in the kernel text section, which is made
read-only after boot (see commit f4f8398 in Cros K3.10 "CHROMIUM:
ARM: mm: fix section mappings after boot").

BUG=chrome-os-partner:31211
TEST=suspend/resume from LP0;
     compiles ok in both arm/arm64 build env;
     boot on norrin64

Change-Id: I084fdbd5b9421c183b0a693ffc909fd074b80a76
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Signed-off-by: Tuomas Tynkkynen <ttynkkynen@nvidia.com>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/215448
Reviewed-by: David Riley <davidriley@chromium.org>