CHROMIUM: soc/tegra: add suspend/resume for TSC
The clock source of the system counter (TSC) is not available when the
system suspends to deep sleep mode. It needs to be switched to the 32KHz
before suspending and restored after resuming.

The Tegra124/132 needs a SW WAR to avoid PMC wake status getting reset
when clearing TSC_MULT_EN in TSC resume code. To fix this, the ENABLE bit
also needs to be cleared at the same time.

BUG=chrome-os-partner:31211
TEST=suspend/resume form LP0;
     compiles ok in both arm/arm64 build env;
     boot on norrin64

Change-Id: I760676f7719d1bee1c9ae8cff5918f66eb2c7054
Signed-off-by: Joseph Lo <josephl@nvidia.com>
[ttynkkynen: squashed in the Tegra124 WAR]
Signed-off-by: Tuomas Tynkkynen <ttynkkynen@nvidia.com>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/215449
Reviewed-by: David Riley <davidriley@chromium.org>