FROMLIST: of: Add nvidia,controller-id property to Tegra I2C bindings
Sometimes, hardware blocks want to issue requests to devices
connected to I2C buses by itself. In such case, the bus the
target device resides on must be configured into a register.
For this purpose, each I2C controller has a defined ID known
by the hardware. Add a property for these IDs to the device tree
bindings, so that drivers can know what ID to write to a hardware
register when configuring a block that sends I2C messages autonomously.

This patch was posted to:
https://lkml.org/lkml/2014/8/13/375

BUG=chrome-os-partner:30834
TEST=None

Change-Id: Ia39f8304e99241ba05ab00ace40df738859c0690
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/212437
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>