dm: allow active and inactive tables to share dm_devs
When a new table is created, in needs to open all of the devices that it
will use.  For multipath, this can cause problems because it may need to
reload a table while some devices are unusable (failed). To get around
this, device mapper now stores the dm_devs list for all table devices in
the mapped_device structure, and the tables' device lists just point to
this main table_devices list.

The device list code in dm.c isn't a straight copy/paste from the code in
dm-table.c, but it's very close (aside from some variable renames).  One
subtle difference is that find_table_device for the tables_devices list
will only match devices with the same name and mode.  This is because we
don't want to upgrade a device's mode in the active table when an
inactive table is loaded.

Access to the mapped_device structure's tables_devices list requires a
mutex (tables_devices_lock), so that tables cannot be created and
destroyed concurrently.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>