CHROMIUM: ARM64: DTS: tegra132: Add PMC thermtrip programming to Norrin
This adds the required information to reset the board during an overheating
situation to the Jetson TK1 device tree. The thermal reset is handled by the
PMC by sending an I2C message to the PMIC. The entries specify the I2C
message to be sent.

BUG=chrome-os-partner:30834
TEST=hardcode the thermtrip to lower temp in DT, the system will be
shutdown when touch the thermtrip

Change-Id: If066b84e6c0f0891fdee212b7496d4e97dc93e7d
Signed-off-by: Wei Ni <wni@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/212441
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>