FROMLIST: ARM: rockchip: rk3288: Switch to use the proper PWM IP
The rk3288 SoC has an option to switch all of the PWMs in the system
between the old IP block and the new IP block.  The rk3288 PWM driver
is written for the new block, so make sure that we enable the new
block in the GRF (general register file) when the PWM driver probes.

We emulate the solution to this problem used in i2c-rk3x.c.  One
difference for the PWM is that there's a single "enable new IP" that's
used for all 4 PWMs.  That means we set this bit 4 times if we've got
all 4 PWMs enabled.  That shouldn't hurt.

BUG=chrome-os-partner:30167
TEST=With series, backlight works:
  cd /sys/class/backlight/backlight*/
  for i in $(seq 255); do echo $i > brightness; sleep .01; done

Change-Id: I5ae7b8d13b17c130723f0048abe26e1e15d8c69e
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/4753221/)
Reviewed-on: https://chromium-review.googlesource.com/212771
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>