FROMLIST: clk: tegra: Add binding for the Tegra124 DFLL clocksource
The DFLL is the main clocksource for the fast CPU cluster on Tegra124
and also provides automatic CPU rail voltage scaling as well. The DFLL
is a separate IP block from the usual Tegra124 clock-and-reset
controller, so it gets its own node in the device tree.

Signed-off-by: Tuomas Tynkkynen <ttynkkynen@nvidia.com>
Signed-off-by: Mikko Perttunen <mikko.perttunen@kapsi.fi>
Acked-by: Michael Turquette <mturquette@linaro.org>
(am from https://patchwork.kernel.org/patch/6397911/)

BUG=chrome-os-partner:40279
TEST=None

Change-Id: Ia7d03e9a52e3bfbceb05a4762e85ba6db80e1f87
Signed-off-by: Penny Chiu <pchiu@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/271397
Reviewed-by: Benson Leung <bleung@chromium.org>