CHROMIUM: clk: tegra: add DFLL dynamic tuning support
This patch is heavily based on the DFLL driver which is
written by Paul Walmsley and used by Tegra124 Chrome
projects currently.

The DFLL output can be fine tuned by the tuning parameters
depending on the voltage threshold. Both of the tuning
parameters and voltage threshold are specified by the
chip characterization. This patch introduces a tuning
timer which controls/monitors the tuning status. When
the DFLL is in open loop, only the tune low parameter
can be used. Once the DFLL enters the close loop, we
can switch to the tune high state if the output voltage
goes higher than the predefined threshold.

BUG=chrome-os-partner:31429
TEST=Checked tune0/tune1 registers with different
frequencies/voltages on Tegra132 Norrin

Change-Id: I135017a1e9a723bb4d8ca9c25898f1b42e6b1305
Signed-off-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/216235
Reviewed-by: David Riley <davidriley@chromium.org>