CHROMIUM: drivers: net: Add nss-gmac driver
This driver adds support for the internal GMACs on IPQ806x SoCs.
It relies (and depends) on nss-drv to load the firmware and interface
with it, but will register the network devices in the system.
It supports the device-tree and will register up to 4 ethernet
interfaces.

This driver is originated from the following location:
*git clone git://codeaurora.org/quic/qsdk/oss/lklm/nss-gmac
*git checkout origin/banana_w14

BUG=chrome-os-partner:31899
TEST=Build & run on the storm platform. This change will do nothing
until nss-drv is enabled in the config and added to the DT

Change-Id: I4946f081c40108138244fd271764c93212ad1f46
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/216832
Tested-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Queue: Bryan Freed <bfreed@chromium.org>