CHROMIUM: MFD: RK808: Add init_ack_masked
The rk808 interrupt mask registers are not truly an interrupt mask
registers in the pure sense.  They are actually interrupt enable
registers.  Specifically, that means:

- If an interrupt is fired and you "mask" it, it's still asserted
  until it is acked.
- If an interrupt is "masked" and the event goes off (like an RTC
  alarm firing), you won't see the event when the interrupt is
  "unmasked".

It appears that regmap-irq is designed for this.  If you set
"init_ack_masked" to true then it will handle acking interrupts when
they are masked.

This is expected to be be squashed into (MFD: RK808: Add new mfd
driver for RK808) by Chris.  Once everything has landed we'll pick
up the official series.

BUG=chrome-os-partner:30167
TEST=Can now boot with proper DTS and use wakealarm.

Change-Id: I6b22fde2f43993c810d326caf8ee7806d5954120
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214241
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>