UPSTREAM: UBI: Dispatch update notification if the volume is updated
The UBI_IOCVOLUP ioctl is used to start an update and also to
truncate a volume. In the first case, a "volume updated" notification
is dispatched when the update is done.

This commit adds the "volume updated" notification to be also sent when
the volume is truncated. This is required for UBI block and gluebi to get
notified about the new volume size.

TEST=tested with other ubi/ubifs fixes as described in http://go/nand-feeding
BRANCH=none
BUG=chromium:441434

Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Cc: stable@vger.kernel.org # v3.15+
(cherry picked from commit fda322a1b3b9e8ee231913c500f73c6988b1aff5)
Signed-off-by: Dan Ehrenberg <dehrenberg@chromium.org>

Change-Id: I3e25db3541914ee87a965687ec9e5f9bea6b68fd
Reviewed-on: https://chromium-review.googlesource.com/234775
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Daniel Ehrenberg <dehrenberg@chromium.org>
Tested-by: Daniel Ehrenberg <dehrenberg@chromium.org>