HACK: ipq806x: uncompress kernel @ 0x42208000
On multiplatform kernel, the memory base address is calculated
automatically. Unfortunately, this results in the kernel being located
at the beginning of the RAM space.

On IPQ platforms, as the NSS core can only access the low memory
addresses, we need to reserve some of this memory.

As we cannot force the decompressor to free-up a certain RAM space, we
will disable the automatic address calculation, and force it to
0x42208000.

BUG=chrome-os-partner:31693
TEST=booted storm platform with 3.14 kernel and device tree at 0x43000000

Change-Id: I47c48aa1d672bb4fd183143b64d659944a2ad256
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/218602
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>