CHROMIUM: iwl7000: disable U-APSD by default
With a significant number of deployed APs, enabling U-APSD leads to the
AP never using aggregation sessions (likely due to the complexities
involved in handling U-APSD in those.) This obviously results in a large
drop in throughput with such APs.

On the other hand, U-APSD can result in some power consumption benefits,
but for now just disable it to get performance with affected APs back
up.

BUG=chrome-os-partner:31397
TEST=connect to Aruba network and check download speed

Change-Id: Ie1fd35f660a33f323394d3afa699e5a5c7185234
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/215797
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>