CHROMIUM: wireless-3.xx: Move uapi include in USERINCLUDE
The include directories in include/wireless-3.xx have to be parsed in a
particular order by the compiler:
*1) include/wireless-3.xx
*2) include
*3) include/wireless-3.xx/uapi
*4) include/uapi

The parsing order with the current code is 1) 3) 2) 4), which works
apparently just fine for 3.8.
However, in 3.16, the file if_ether.h doesn't exist anymore in
include/linux. Therefore, that leads to the following situation:
*legacy drivers (outside of the wireless-3.16 directories) which use
 this file end up including include/wireless-3.16/uapi/linux/if_ether.h,
 which causes all kind of build issues.
*ported drivers will use a #include <uapi/linux/if_ether.h, which will
 include the alternative file from the corresponding stack.

The new include order works for 3.16, and should not impact other stack.

BUG=chrome-os-partner:31738
TEST=Build and successfully brought up Storm WiFi on a local build
including Storm changes

Change-Id: I939d06e0220568be2cd41dc148c15e22f0313c6a
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/215999
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>
Tested-by: Toshi Kikuchi <toshik@chromium.org>
Commit-Queue: Toshi Kikuchi <toshik@chromium.org>