UPSTREAM: mfd: RK808: Add new mfd driver for RK808
The RK808 chip is a power management IC for multimedia and handheld
devices. It contains the following components:

- Regulators
- RTC
- Clkout

The RK808 core driver is registered as a platform driver and provides
communication through I2C with the host device for the different
components.

Conflicts:
	drivers/mfd/Kconfig
	drivers/mfd/Makefile
...trivial conflicts because we don't have MFD_RN5T618 in our tree.

BUG=chrome-os-partner:30167
TEST=rk808 working

Change-Id: I452a643bcb1f0f2c12cc7b74b671d261eb17c15b
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Signed-off-by: Zhang Qing <zhangqing@rock-chips.com>
Tested-by: Heiko <heiko@sntech.de>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org lee/mfd.git for-mfd-next
 commit 28d85ec98384251a91e7123f22432660a272962f)
Reviewed-on: https://chromium-review.googlesource.com/216389