CHROMIUM: cyapa: Add support for probing device from ACPI
With firmware support it is possible to discover and probe devices on I2C buses without needing the glue layer in chromeos_laptop that we use today.

This requires that the drivers implement the new acpi_match_table with compatible ACPI IDs.

I chose to use CYAP0000 for the cyapa touchpad as a generic HID that can be used in ACPI tables to probe a CYAPA touchpad.

This is based on dlaurie's patch from:
https://chromium-review.googlesource.com/#/c/182340/

BUG=chrome-os-partner:31494
TEST=Cypress touchpad is working on Auron

Change-Id: I943f24c9a4450b552cd020b109afb720081f84a7
Reviewed-on: https://chromium-review.googlesource.com/216362
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Tested-by: Wenkai Du <wenkai.du@intel.com>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>