FROMLIST: dts, arm64: Add dtbs_install make target
This adds the dtbs_install make target to arm64. The target has been
introduced already to arch/arm with the following commit:

 f4d4ffc03efc kbuild: dtbs_install: new make target

Implementation for arm64 is the same as for arm.

With 'dtbs_install' all config enabled dtb files are installed to
either the INSTALL_DTBS_PATH directory or the default location:

 $INSTALL_PATH/dtbs/$KERNELRELEASE

Signed-off-by: Robert Richter <rrichter@cavium.com>
(am from https://patchwork.kernel.org/patch/4850001/)

BUG=chrome-os-partner:30840
TEST=Build for Ryu and Veyron

Change-Id: I1d1aa4dbbc7d6c8758fa630ab43e707d89f26b74
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/218372
Reviewed-by: Doug Anderson <dianders@chromium.org>