HACK: ARM: dts: Temporary change to use OTP_OUT as a GPIO reset
We've decided that the best way to reboot our board is to reuse the
"over temperature" pin.  This will not just reboot the SoC but will
also notify other components on the board that a restart has happened.

Eventually we'll want to change OTP_OUT to be controlled by the
thermal driver.  ...but that's not ready yet.  For now, we can just
enable the pin as a GPIO.

BUG=chrome-os-partner:31605
TEST=reboot works and resets the TPM now

Change-Id: I8801853c351a9ef8c3a0054b8cbdbc27e12db129
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217076
Reviewed-by: David Riley <davidriley@chromium.org>