UPSTREAM: usb: dwc2/gadget: fix phy initialization sequence
In the Generic PHY Framework a NULL phy is considered to be a valid phy
thus the "if (hsotg->phy)" check does not give us the information whether
the Generic PHY Framework is used.

In addition to the above this patch also removes phy_init from probe and
phy_exit from remove. This is not necessary when init/exit is done in the
s3c_hsotg_phy_enable/disable functions.

BUG=chrome-os-partner:32735
TEST=Can use dwusb2 ports

Change-Id: Idbc0c54fa7cc022afad00526b6e5843503ebb0a0
Signed-off-by: Kamil Debski <k.debski@samsung.com>
Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Robert Baldyga <r.baldyga@samsung.com>
Cc: stable <stable@vger.kernel.org> # 3.16
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>
(cherry picked from linuxnext
 commit ca2c5ba80f823a38c4aae506e2f03c1afb076b0a)
Reviewed-on: https://chromium-review.googlesource.com/222270
Reviewed-by: Julius Werner <jwerner@chromium.org>