mm: zone_page_state(): regularize UP and SMP behavior
zone_page_state is an API hazard because of the difference in behaviour
between SMP and UP is very surprising.  There is a good reason to allow
NR_ALLOC_BATCH to go negative -- when the counter is reset the negative
value takes recent activity into account.  This patch makes
zone_page_state behave the same on SMP and UP as saving one branch on UP
is not likely to make a measurable performance difference.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Reported-by: Vlastimil Babka <vbabka@suse.cz>
Reported-by: Leon Romanovsky <leon@leon.nu>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>