UPSTREAM: PM / AVS: rockchip-io: add driver handling Rockchip io domains
IO domain voltages on some Rockchip SoCs are variable but need to be
kept in sync between the regulators and the SoC using a special
register.

A specific example using rk3288:
- If the regulator hooked up to a pin like SDMMC0_VDD is 3.3V then
  bit 7 of GRF_IO_VSEL needs to be 0.  If the regulator hooked up to
  that same pin is 1.8V then bit 7 of GRF_IO_VSEL needs to be 1.

Said another way, this driver simply handles keeping bits in the SoC's
general register file (GRF) in sync with the actual value of a voltage
hooked up to the pins.

Note that this driver specifically doesn't include:
- any logic for deciding what voltage we should set regulators to
- any logic for deciding whether regulators (or internal SoC blocks)
  should have power or not have power

If there were some other software that had the smarts of making
decisions about regulators, it would work in conjunction with this
driver.  When that other software adjusted a regulator's voltage then
this driver would handle telling the SoC about it.  A good example is
vqmmc for SD.  In that case the dw_mmc driver simply is told about a
regulator.  It changes the regulator between 3.3V and 1.8V at the
right time.  This driver notices the change and makes sure that the
SoC is on the same page.

CQ-DEPEND=CL:I7c04f7706667f79f266d646ea2f9b6c718370e52
BUG=chrome-os-partner:30167
TEST=Build and boot

Change-Id: I64ca561bb951b652161890d41b450083f41968fe
Signed-off-by: Heiko Stübner <heiko@sntech.de>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
[khilman: fix compiler warnings]
Signed-off-by: Kevin Hilman <khilman@linaro.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org khilman/linux.git avs-next
 commit 662a958638bdb802451da6933e7ddb4a69517893)
Reviewed-on: https://chromium-review.googlesource.com/220685
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>