[media] BZ#84401: Revert "[media] v4l: vb2: Don't return POLLERR during transient buffer underruns"
This reverts commit 9241650d62f79a3da01f1d5e8ebd195083330b75.

The commit 9241650d62f7 was meant to solve a race issue that
affects Gstreamer version 0.10, when streaming for a long time.

It does that by returning POLERR if VB2 is not streaming.

However, it broke VBI userspace support on alevt and mtt (and maybe
other VBI apps), as they rely on the old behavior.

Due to that, we need to roll back and restore the previous behavior.

For more details, see:
	https://bugzilla.kernel.org/show_bug.cgi?id=84401

So, let's rollback the change for now, and work on some other
fix for it.

Cc: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Cc: Hans Verkuil <hans.verkuil@cisco.com>
Cc: Pawel Osciak <pawel@osciak.com>
Cc: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>