md: Make sync_page_io aware of underlying block size.
sync_page_io is sometimes called on to read an object smaller than a
page.  If that happens to be smaller than the physical block size
of the device, the size needs to be rounded up.

This is particularly import for use of sync_page_io by dm-raid.

This is a simple fix for a bug that stops dm-raid from working on 4K
phys-sector devices.  So it seems suitable for -stable at low
priority.

Signed-off-by: NeilBrown  <neilb@suse.de>
Cc: stable@vger.kernel.org (After release of 3.18)
Reported-by: Liuhua Wang <lwang@suse.com>