UPSTREAM: cfg80211: add wowlan net-detect support
(upstream commit 8cd4d4563ef0a518002c4a8f47dd950afe386ea8)

Add a new WoWLAN API to enable net-detect as a wake up trigger.
Net-detect allows the device to scan in the background while the
host is asleep to wake up the host system when a matching network
is found.

Reuse the scheduled scan attributes to specify how the scan is
performed while suspended and the matches that will trigger a
wake event.

BUG=N/A
TEST=compile iwl7000 driver using this infrastructure

Change-Id: Ifb80ee12dff845720ed0c6fb583a9aee6e309490
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/273858
Reviewed-by: Paul Stewart <pstew@chromium.org>