CHROMIUM: driver-core: add test module for asynchronous probing
This test module tries to test asynchronous driver probing by having a
driver that sleeps for extended period of time (5 secs) in its probe()
method, and measuring time needed to register this driver (with device
already registered) and a new device (with driver already registered).
The module will fail to load if time spent in register call is more than
half the probing timeout.

As a sanity check the driver than will try to synchronously register
driver and device and fail if registration will take less than half
timeout time.

BUG=chromium:415343
TEST=Compiled and tried loading the driver

Change-Id: I02fcb03221bcb4c46e1116f9c49fb37c8d32a917
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219107
Reviewed-by: Olof Johansson <olofj@chromium.org>