ipc: resolve shadow warnings
Resolve some shadow warnings produced in W=2 builds by changing the name
of some parameters and local variables.  Change instances of "s64" because
that clashes with the well-known typedef.  Also change a local variable
with the name "up" because that clashes with the name of of the "up"
function for semaphores.  These are hazards so eliminate the hazards by
renaming them.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>