CHROMIUM: MIPS: Add support for IMG Pistachio SoC
Add initial support for boards based on the Imagination Pistachio SoC.
Pistachio is based on a dual-core MIPS interAptiv CPU and will boot
using device-tree.

The only currently supported platform is the Pistachio FPGA, which will
use a built-in device-tree and lacks several hardware blocks found in
the actual silicon.

BUG=chrome-os-partner:30840
TEST=Build and boot on FPGA

Change-Id: Iac907729eaf7c169fe3493a6225fbacc57ea08da
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220066
Reviewed-by: James Hartley <james.hartley@imgtec.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>