dm raid: add discard support for RAID levels 4, 5 and 6
In case of RAID levels 4, 5 and 6 we have to verify each RAID members'
ability to zero data on discards to avoid stripe data corruption -- if
discard_zeroes_data is not set for each RAID member discard support must
be disabled.

Also add an 'ignore_discard' table argument to the target in order to
ignore discard processing completely on a RAID array, hence not passing
down discards to MD personalities.

This 'ignore_discard' control provides the ability to:
- prohibit discards in case of _potential_ data corruptions in RAID4/5/6
  (e.g. if ability to zero data on discard is flawed in a RAID member)
- avoid discard processing overhead

Signed-off-by: Heinz Mauelshagen <heinzm@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>