ARM: dts: Fix chip select GPIO on exynos5250-smdk5250
Commit 65bbe3fee0e1 ("ARM: dts: Clean up exynos5250-smdk5250")
improved the smdk5250 DTS but the patch was rebased due conflicting
changes and the merge resolution added a regression of the bug fixed
in commit e138d4333aa0 ("ARM: dts: fix the chip select gpios definition
in the SPI nodes").

This patch fixes the issue by removing the old cs-gpio and using
the generic cs-gpios property.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>
Signed-off-by: Kukjin Kim <kgene.kim@samsung.com>