mtd: msm_nand: fix ecc err on reading erased pgs with bit-flips
Read of erased pages with bit-flips in msm_nand_read_oob caused
it to report uncorrectable errors to its caller. In the fix, when
uncorrectable errors are reported by the Zoe controller for a page,
the page is read again in raw mode & is compared with an all 0xFF
data. If the number of bit-flips in a code-word exceeds supported
ecc strength then -EBADMMSG error is returned to the caller else
-EUCLEAN is returned to the caller after fixing all flips in the
buffer passed.

Change-Id: Iebf413e7b32d36c8caba826c0f5d46c22653e246
Acked-by: John Nicholas <jnichola@qti.qualcomm.com>
Signed-off-by: Venkatesh Yadav Abbarapu <quicvenkat@codeaurora.org>