storm: config: Move the TPM driver from the kernel image to a module.
This is a workaround to a recently discovered timing issue.
The I2C adapter fails transactions with EIO while the kernel is booting,
but it works once the kernel is up.
Moving the TPM driver out of the kernel image to a loadable module
gives enough of a delay for this to work.

BUG=chrome-os-partner:32378
TEST=Boot the image, then check for /dev/tpm0.

Change-Id: Ie217b8daac70cf33b58e1c85e07950551260ed48
Signed-off-by: Bryan Freed <bfreed@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220200
Reviewed-by: Olof Johansson <olofj@chromium.org>