FIXUP: UPSTREAM: clk: rockchip: add new clock-type for the cpuclk
This is a squash of a revert of:
  fb8f9d2 FROMLIST: clk: rockchip: add new clock-type for the cpuclk
...and an upstream cherry pick to match what landed upstream.

---

When changing the armclk on Rockchip SoCs it is supposed to be reparented
to an alternate parent before changing the underlying pll and back after
the change. Additionally there exist clocks that are very tightly bound to
the armclk whose divider values are set according to the armclk rate.

Add a special clock-type to handle all that. The rate table and divider
values will be supplied from the soc-specific clock controllers.

CQ-DEPEND=CL:I1638c3f8e03a6f164bfc8c8f394c99387aba10a9
BUG=chrome-os-partner:32090
TEST=w/ series, run stress test from chrome-os-partner:32090

Change-Id: I3b0c396a31f412c1c68eb7299d40432f11840aec
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Reviewed-by: Doug Anderson <dianders@chromium.org>
On a rk3288-board:
Tested-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.linaro.org/people/mike.turquette/linux.git clk-next
 commit f6fba5f6967dbc062a7c138d67e2314220f5dd04)
Reviewed-on: https://chromium-review.googlesource.com/220681
Reviewed-by: Derek Basehore <dbasehore@chromium.org>