CHROMIUM: Mali: rockchip: fetch freq/voltage table from dts
Extract freq and voltage from .dts "operating-points" field.   The field
is formatted per [0].  However, we are not using these values with cpufreq,
nor are we using the rest of the opp infrastructure.  So, instead of using
of_init_opp_table() and dev_pm_opp_init_cpufreq_table(), we just manually
extract the frequency and voltage and save them into a device frequency
and voltage table.

[0] https://www.kernel.org/doc/Documentation/devicetree/bindings/power/opp.txt

The operating points will be used by the DVFS algorithm to set the GPU
frequency and voltage at runtime.

For now, though, we just use the lowest frequency value to initialize
the gpu frequency.  A future patch will also initialize the gpu regulator
to the corresponding lowest voltage.

BUG=chrome-os-partner:31652
TEST=emerge-veyron_pinky chromeos-kernel-3_14
TEST=boot to UI with DRM patch set

Change-Id: Iee9ab41fe4cea6392826e28d93a862b755549817
Signed-off-by: Simon Xue <xxm@rock-chips.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220079
Reviewed-by: Dominik Behr <dbehr@chromium.org>