[media] si2165: do load firmware without extra header
The new file has a different name: dvb-demod-si2165-d.fw

Count blocks instead of reading count from extra header.
Calculate CRC during upload and compare result to what chip calcuated.
Use 0x01 instead of real patch version, because this is only used to
check if something was uploaded but not to check the version of it.

[mchehab@osg.samsung.com: fix a trivial merge conflict and the
 MODULE_FIRMWARE macro]
Signed-off-by: Matthias Schwarzott <zzam@gentoo.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>