BACKPORT: gpu: host1x: mipi: Preserve the contents of MIPI_CAL_CTRL
By paving the CTRL reg value, the current code changes MIPI_CAL_PRESCALE
("Auto-cal calibration step prescale") from 1us to 0.1us (val=0). In the
description for PHY's noise filter (MIPI_CAL_NOISE_FLT), the TRM states
that if the value of the prescale is 0 (or 0.1us), the filter should be
set between 2-5. However, the current code sets it to 0.

For now, let's keep the prescale and filter values as-is, which is most
likely the power-on-reset values of 0x2 and 0xa, respectively.

BUG=chrome-os-partner:30799
TEST=Tested on ryu

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit eb83160f0308253d668a5f3b99e0504d4f54b1ff)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Change-Id: Id276daae671c4b9ed2e725f8d90e7dafac293340