Revert "BACKPORT: ASoC: Intel: Test Fix for suspend/resume"
This reverts commit b1bb49da0b5ac6e50b2f35573538c66c59595c7d.

Conflicts:
	sound/soc/intel/sst-baytrail-dsp.c
	sound/soc/intel/sst-baytrail-ipc.c

Reverting in favor of upstream solution.

The conflict in sst-baytrail-ipc.c is a result of commit
aecff8e896edefdd6d72965ee7e851b1c95f81f0.  The revert
operation didn't know what to do with sst_byt_register_notifier.
There does not appear to be any upstream equivalent to this, but
left it preserved for now.

The conflict in sst-baytrail-dsp.c is a result of commit
8fdf576d6f8cd450dcee7bfa4ca962e6caab1a4c.  The revert operation
was trying to reintroduce the code that saves the physical address
of extended firmware block in sst_byt_init that got relocated to
sst_byt_boot.  Preserved the code relocation as it matches
upstream.

BUG=chrome-os-partner:28651, chromium:351000
TEST=With patch series applied...
1) Play audio
   Close lid ~10 seconds
   Open lid
   System and audio resumes
2) Play audio
   $ echo none > /sys/power/pm_test
   $ powerd_dbus_suspend
   Tap spacebar after ~10 seconds
   System and audio resumes
3) Play audio
   $ echo devices > /sys/power/pm_test
   $ powerd_dbus_suspend
   System and audio resumes after ~5 seconds

Change-Id: I1089a9bd035ffed68d58c73918a8806931111538
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/219483
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Matthew E Hiltner <matthew.e.hiltner@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/225549
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>