FROMLIST: ARM: rockchip: convert to regmap and use pmu syscon if available
The pmu register space is - like the GRF - shared by quite some peripherals.
On the rk3188 and rk3288 even parts of the pinctrl are living there.
Therefore we normally shouldn't map it a second time when the syscon
does this already.

Therefore convert the cpu power-domain handling to access the pmu via a
regmap and at first try to get it via the syscon interface.
Getting this syscon will only fail if the pmu node does not have the
"syscon" compatible and thus does not get shared with other drivers.

In this case we map it like before and create the necessary regmap on
top of it.

BUG=chrome-os-partner:30167
TEST=Boot and can still use all 4 CPUs (with series)

Change-Id: I24efe241f28739eb2010a2899bdb0de1c71f555d
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5077241/)
Reviewed-on: https://chromium-review.googlesource.com/223334
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>