FROMLIST: mmc: dw_mmc: rockchip: Don't recalc the clock when it goes off
The "set_ios" function is called with a clock of 0 when the clock is
turning off.  There's no reason to go through all the extra Rockchip
logic (whose goal is to make sure DIV is 0 or 1) in that case.  The
Rockchip logic happened to work because the CCF will pick the lowest
possible rate when you ask it for a clock of 0, but it's silly to go
through all the remuxing and adjusting for no reason.

BUG=chrome-os-partner:30167
TEST=Build and boot

Change-Id: Ie17c2c3756f55a9961e529a6b5773047f7c1812c
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5081011/)
Reviewed-on: https://chromium-review.googlesource.com/223264
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>