FROMLIST: drivers: cpuidle: add status property to ARM idle states
On some platforms the device tree bindings must provide the kernel
with a status flag for idle states, that defines whether the idle
state is operational or not in the current configuration.

This patch adds a status property to the ARM idle states compliant
with ePAPR v1.1 and updates the DT parsing code accordingly.

Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Acked-by: Kevin Hilman <khilman@linaro.org>
(am from https://patchwork.kernel.org/patch/5092151/)

BUG=chrome-os-partner:32136
TEST=boot test on Ryu

Change-Id: I1de684c49ef113265bb5653aca1b5301c3e0f616
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/226756
Reviewed-by: Aaron Durbin <adurbin@chromium.org>