FROMLIST: drivers: cpuidle: add idle-state-name description to ARM idle states
On ARM machines, where generally speaking the idle state numbering has
no fixed and standard meaning it is useful to provide a description
of the idle state inner workings for benchmarking and monitoring purposes.

This patch adds a property to the idle states bindings that if present
gives platform firmware a means of describing the idle state and export
the string description to user space.

The patch updates the DT parsing code accordingly to take the description,
if present, into consideration.

Cc: Kevin Hilman <khilman@linaro.org>
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
(am from https://patchwork.kernel.org/patch/5092161/)

BUG=chrome-os-partner:32136
TEST=boot test on Ryu

Change-Id: If10dc454396ae8afd098eec2cb03a511a3d3cbd5
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/226757
Reviewed-by: Aaron Durbin <adurbin@chromium.org>