CHROMIUM: usb: don't resume root hub async in dark resume
Some async work is scheduled to resume root hubs, but if we are
keeping the usb ports off during dark resume, there is no reason
for this work. Furthermore, when the work is scheduled, a flag is
set that is not cleared until the work is completed. Most of the
time this flag is cleared early enough in dark resume to let the
device resuspend correctly, but if the order in which the devices
are suspended changes, we might fail to resuspend and sit in S0
for an extra ten seconds while we wait for the work to finish and
the bit to clear.

This change prevents the work from being scheduled in the first
place.

BUG=chrome-os-partner:31811
TEST=As root:
  echo enabled > /sys/bus/usb/devices/usb1/power/dark_resume_active
  echo enabled > /sys/bus/usb/devices/usb2/power/dark_resume_active
  echo 0.0 15 > /var/lib/power_manager/dark_resume_suspend_durations
  restart powerd
  powerd_dbus_suspend
and verify that the system performs dark suspend/resume cycles with
and without USB devices plugged in.

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Change-Id: I6a39986bb4278af22b93fd38c4024c2dd903ac7b
Reviewed-on: https://chromium-review.googlesource.com/224450
Reviewed-by: Julius Werner <jwerner@chromium.org>