CHROMIUM: drm: rockchip: fix rockchip_drm_find_crtc
The variable i is used un-initialized.  I am not sure why gcc can't detect
this.

Note: This fix is compatible with the upstream version of drm/rockchip.  A
simpler solution is to compare pipe with crtc->index, but that depends
on the atomic patchset (which introduces crtc->index).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:399935
TEST=Boot to UI on veyron_pinky

Change-Id: Ie7b4f8da86dce145f7e89a351f517108bfd867b0
Reviewed-on: https://chromium-review.googlesource.com/224328
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>