ARM: dts: add dts file for rk3288-jerry-rev1
Moved rk3288-pinky-rev2 content to rk3288-veyron.dtsi as pinky
derivatives will be very close to this.

Added rk3288-jerry-rev1 as essentially a copy of
rk3288-pinky-rev2 for now.

rk3288-jerry-rev1 was used instead of rk3288-jerry-rev0 so that
it matches the same revision scheme as pinky, which essentially says:
GPIO: 0000 -> rev 1
GPIO: 0001 -> rev 2
so to get a real jerry-rev0 we would need to have two different
tables

BUG=chrome-os-partner:33269
TEST=build successful, boot on pinky rev2 with fake
FW and boots

Change-Id: I43eebcd215573b52de739f4055967d41930991a3
Signed-off-by: Katie Roberts-Hoffman <katierh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/225293
Reviewed-by: Doug Anderson <dianders@chromium.org>