xen/arm: introduce xen_is_dma_coherent
Introduce a utility function to check whether a device is dma coherent.
The current implementation is suboptimal. Let's rework it later in
common arch/arm code.

Use it to figure out whether we need to issue cache maintenance
operations instead of checking on the existence of a particular dma_ops
function for the device.

This is correct because arm_coherent_dma_ops is the dma_map_ops for
coherent devices, which do not need cache maintenance operations. In
fact arm_coherent_dma_ops does not set the hooks that we were
previously checking for existance.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>

---

Changes in v5:
- introduce xen_is_dma_coherent as a xen specific function;
- call xen_is_dma_coherent instead of is_dma_coherent.