UPSTREAM: ASoC: dapm: Do not add un-muxed paths to MUX control
Paths that are directly connected to a MUX widget are not affected by
changes to the MUX's control. Rather than checking if a path is directly
connected each time the MUX is updated do it only once when MUX is created.

We can also remove the check for e->texts[mux] != NULL, since if that
condition was true the code would have had already crashed much earlier (And
generally speaking if a enum's 'texts' entry is NULL it's a bug in the
driver).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 98407efc1384b31cdcb1eeddc74ee35499d3418f)
BUG=chrome-os-partner:40036
TEST=Audio playback works on test device

Change-Id: I1da4264d9c2d26ad5574bb904a5d31b4fe85c075
Reviewed-on: https://chromium-review.googlesource.com/270848
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>