CHROMIUM: iommu/rockchip: make use of iommu groups
The following upstream commit made the .add_device callback mandatory, or
else an iommu driver will not get installed on a bus:

commit fb3e306515ba6a012364b698b8ca71c337424ed3
Author: Mark Salter <msalter@redhat.com>
Date:   Sun Sep 21 13:58:24 2014 -0400

    iommu: Fix bus notifier breakage

So, to stay compatible with upstream, we create an add_device() callback.

This callback creates an "iommu group" for any platform device that is
also an iommu master, as denoted by having an "iommus" property in its
devicetree node.

The add_device() callback also stores the iommu device in the iommu group.
This means that iommu master drivers no longer need to manually parse
devicetree and set archdata.iommu itself.
Hence, we remove vop_iommu_init().

Finally, in v3.14, dev_get_drvdata() would happily return NULL if dev=NULL.
This "feature" was removed in v3.15 by the following patch, causing
dev_get_drvdata() to crash when given a NULL dev.

commit d4332013919aa87dbdede67d677e4cf2cd32e898
Author: Jean Delvare <jdelvare@suse.de>
Date:   Mon Apr 14 12:57:43 2014 +0200

    driver core: dev_get_drvdata: Don't check for NULL dev

To fix this issue in attach and detach, we return immediately if the
device does not have an iommu group, rather than relying on
dev_get_drvdata() to return NULL if the given device was NULL.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:31704
TEST=sanity check UI on pinky device

Change-Id: I37fe7bc5165feee0badf25137cd8d8cdab54b98b
Reviewed-on: https://chromium-review.googlesource.com/225594
Reviewed-by: 智情 姚 <mark.yao@rock-chips.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>