CHROMIUM: drm: fix edid update on unplug
After destroying the edid blob we must clear pointer to it.  Otherwise
it still points to the old cleared blob object in the !edid case.
This causes drm_mode_connector_update_edid_property() to try to destroy it
again later which causes an idr_remove WARN, and an OOPS when
drm_property_destroy_blob() tries to remove it again from the list.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:428216
TEST=On pinky w/ HDMI patchset, hotplug; unplug => no crash

Change-Id: I59ef1b184d4f50474ac4edb5c66fa4ecbaf23431
Reviewed-on: https://chromium-review.googlesource.com/226512
Reviewed-by: David Sodman <dsodman@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>