UPSTREAM: ath10k: fix pm resume after suspend
Firmware was crashing when we were trying to warm reset it
after suspend. This was due to the fact that target registeres
can be accessed only if the hardware is awaken.

This patch makes sure to awake the device also on the hif up,
not only in case of probe call.

Signed-off-by: Bartosz Markowski <bartosz.markowski@tieto.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>
(cherry picked from commit 707b1bbd7e6ec533cfa027054ccbeb90113a40b0)
(source: https://github.com/kvalo/ath/tree/ath-qca)

BUG=chrome-os-partner:35262
TEST=build/boot on storm suceeded. Setup Storm board as AP using hostapd and
connected to the board using another device. Device was able to connect to
the internet and load multiple websites.

Change-Id: I39a1d07fef2d04082d47af04fbd597b98010476d
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/241606
Reviewed-by: Toshi Kikuchi <toshik@chromium.org>