CHROMIUM: ARM: dts: qcom: add IMEM syscon to IPQ8064
The IPQ8064 has an SRAM block known as IMEM, which is holds various data
about the chip.  In particular, it houses the location whereby the
kernel can access the chip-specific characteristics written during
manufacturing, which will be used by the cpufreq driver to determine
which OPP table to use.

BUG=chrome-os-partner:32317
TEST=Build/boot on storm succeeded

Change-Id: I9d6c35f7ad26566b680d668f02e9a981c9546aaa
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/231799
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Commit-Queue: Viswanath Kraleti <vkraleti@codeaurora.org>
Tested-by: Viswanath Kraleti <vkraleti@codeaurora.org>