CHROMIUM: ACPI / Wakeup: Add wakeup source platform support
This allows ACPI systems to tell what type of wakeup source woke the system. It
uses the ACPI method _SWS (system wake source) in both the _SB and the _GPE
context to find the wakeup source. These are compared to data from acpi wakeup
sources on resume.

BUG=chrome-os-partner:32966
TEST=emerge-samus chromeos-kernel-3_14 succeeds
samus boots and suspend/resumes correctly

Change-Id: I028c83e9b86300034c8f61e72096c6d5ea0e03c8
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226933
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>