drm/i915: Simplify the pfit check in the fastboot hack
Checking for lvds/edp is wrong (since we also support dsi panels now)
and not terribly useful. If we want to be more intelligent about all
this then we need to look at the old state to figure out whether the
pfit adjustment is needed. This goes all the way back to the original
fastboot pfit patch:

commit 4d6a3e63bce0cb604864e36585ca8983160a421a
Author: Jesse Barnes <jbarnes@virtuousgeek.org>
Date:   Wed Jun 26 01:38:18 2013 +0300

    drm/i915: turn off panel fitting at flip time if needed v2

Spotted while reviewing Ander's patch which changed a lot of the
pipe_has_type checks.

Cc: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>