UPSTREAM: mmc: sdhci: Fix ADMA page boundary warnings
Bytes are being copied from/to a single page.  The intent
of the warning is to warn if the page boundary is crossed.
There are two problems.  First, PAGE_MASK is mistaken for
(PAGE_SIZE - 1).  Secondly, instead of using the number
of bytes to copy, the warning is using the maximum that
that value could be.  Fix both.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 8be78c6ad4d38ae555065b2778205d310300c8b0)
Signed-off-by: Marc Herbert <marc.herbert@intel.com>

BUG=chrome-os-partner:40410
TEST=With this whole 64-bit ADMA series applied:
TEST=on a system with a 64-bit ADMA controllers, all initialize with the
 following log message:
 [  103.444685] mmc0: SDHCI controller on PCI [0000:00:1e.4] using ADMA 64-bit
 [  103.495709] mmc1: SDHCI controller on PCI [0000:00:1e.6] using ADMA 64-bit
TEST=run SD/eMMC test plan: no change other than the above.
TEST=same on another system with just 32-bit ADMA: no observable change at all.

Change-Id: I67be1b0bdc6f9c80352d8ae84486a04ef77d7b9e
Reviewed-on: https://chromium-review.googlesource.com/276372
Reviewed-by: Derek Basehore <dbasehore@google.com>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Commit-Queue: Marc Herbert <marc.herbert@intel.com>