drm/i915: Mark fastboot as unsafe
Fastboot in its current incarnation assumes that the pfit isn't
relevant for the state and that it can be disabled without restarting
the crtc. Unfortunately that's not the case on gen2/3 - it upsets the
hw and results in a black screen.

Worse, the way the current fastboot hack is structured we can't detect
and work around this in the code, since the fastboot smashes the
adjusted mode into crtc->mode. Which means the higher levels can't
correctly figure out that this is a lie and act accordingly.

Since fastboot is just a tech demo let's mark the module option as
experimental and close the corresponding reports as wontfix.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=84682
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>