FIXUP: UPSTREAM: clk: rockchip: disable unused clocks
This is a squash of a revert of:
  7f35dc4 FROMLIST: clk: rockchip: disable unused clocks
...and an upstream cherry-pick.  It looks like I jumped the gun just a
few hours and the patch changed just as it landed upstream, so getting
the right version.

The rockchip clock driver use CLK_IGNORE_UNUSED flag to make sure
all the clocks are available like default power on state.
We have implement the clock manage in most of rockchip drivers,
it is time to remove it for power save.
Instead we add CLK_IGNORE_UNUSED for some clock nodes which should
be on during boot or no module driver in kernel will initialize it.

BUG=chrome-os-partner:33187
TEST=See CL:I5f48baad26c3c55bdf8c762f3b9d3c2da3cc71f5

Change-Id: I2edcca6133d869345c45eda2449e6c27bd79bc91
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org mmind/linux-rockchip.git linuxrockchip/v3.19-clk/next
 commit 78eaf6095cc763c1a228ebac5682852f04e85205)
Reviewed-on: https://chromium-review.googlesource.com/227462