FROMLIST: ath10k: read calibration data from Device Tree
V1->V2: fixed a bug that returns an uninitialized value if the node
is not found.

This patch adds support for reading calibration data from Device Tree.
It looks for the calibration data in Device Tree if it can't find it
in a file.

The node for the calibration data should be defined like this:

pci {
        pcie@0 {
                reg = <0 0 0 0 0>;
                #interrupt-cells = <1>;
                #size-cells = <2>;
                #address-cells = <3>;
                device_type = "pci";

                ath10k@0,0 {
                        reg = <0 0 0 0 0>;
                        device_type = "pci";
                        qcom,ath10k-calibration-data = [ 01 02 03 ... ];
                };
        };
};

BUG=chrome-os-partner:33010
TEST=successfully reads "qcom,ath10k-calibration-data" in DT

Change-Id: I6fe3c50fe2de9177d031a4c17bd95152372dd993
Signed-off-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/229385