HACK: CHROMIUM: ath10k: read calibration data in base64 format from DT
Chrome OS firmware doesn't support binary format in VPD currently.
As a workaround, the firmware stores the calibration data in base64
format in the same node with the different name:

  qcom,ath10k-calibration-data-base64 = [ 01 02 03 ... ];

Since the original property "qcom,ath10k-calibration-data" is always
looked for first, it should have an invalid size (e.g. 1).

BUG=chrome-os-partner:33010
TEST=successfully reads "qcom,ath10k-calibration-data-base64" in DT

Change-Id: I5f9281c189b2d8926cbe05975b0e0374c8f775c5
Signed-off-by: Toshi Kikuchi <toshik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/228251
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>