fanotify: ignore marks not respected
The fanotify API allows to mark mounts, directories, and files for
notification. Furthermore it allows to create marks with an ignore mask to
indicate which events shall be ignored.

Function fsnotify_add_inode_mark() stores marks for inodes as list in attribute
i_list of the inode. The list is sorted by priority and group.

Function fsnotify_add_vfsmount_mark() stores marks for mounts as list in
attribute m_list of the vfsmount. The list is sorted by priority and group.

If an event occurs function fsnotify() is called. This function loops over all
mount and inode marks and sends the event to the respective groups using
function send_to_group().

For an event and a group both an inode mark and a vfsmount mark may coexist.
E.g. a user program may want to monitor one mount for FAN_MODIFY events but
wants to ignore FAN_MODIFY events for some file.

If for an event and a group both an inode mark and a vfsmount mark exist, it is
important to call send_to_group() for both marks as once. Otherwise ignore mask
bits set on the one mark cannot be considered for the other mark.

Function fsnotify() assumes that m_list and i_list are sorted by group and
ignores the sorting by priority. Hence some ignore marks may not be correctly
considered.

Bug report 87721 provides an example code demonstrating the error.
https://bugzilla.kernel.org/show_bug.cgi?id=87721

The appended patch changes fsnotify() to assume sorting by priority and group.
Furthermore some comments are added to render the function more legible.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Cc: Jan Kara <jack@suse.cz>
Cc: Eric Paris <eparis@redhat.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>