tty: serial: msm: Reset uartdm after baud rate change
We need to issue a reset if we ever change the value of the IPR
register on DM hardware. If we don't reset the hardware the RX
stale interrupt never triggers and the only way to trigger an RX
handling event is by filling up the fifo. This causes things like
getty to not work so well considering it might change the baud
rate a few times. Fix this by moving the reset on startup and any
reprogramming required after the reset to be after we change the
baud rate.

BUG=chrome-os-partner:32192
TEST=booted the kernel - kernel logs showed up on console, able to
     access the console and issue commands without issue

Change-Id: I435c9caf422140fee2d92320df81103752341c10
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Andy Gross <agross@codeaurora.org>
Signed-off-by: Abhishek Airan <aairan@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/228506
Reviewed-by: <agross@quicinc.com>
Tested-by: <agross@quicinc.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Commit-Queue: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>