UPSTREAM: usb: dwc2: gadget: Do not fail probe if there isn't a clock node
Since the dwc2 hcd driver is currently not looking for a clock node during
init, we should not completely fail if there isn't a clock provided.
By assigning clk = NULL, this allows the driver, when configured for dual-role
mode, to be able to continue loading the host portion of the driver when
a clock node is not specified.

Acked-by: Paul Zimmerman <paulz@synopsys.com>
BUG=chrome-os-partner:33496
TEST=USB still works

Change-Id: Ia1222a7da038b18e60b63c56a296f9d95c2d91db
Signed-off-by: Dinh Nguyen <dinguyen@opensource.altera.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org balbi/usb.git next
 commit 8d736d8a9c44547f14711c52875c88d19d8b287b)

Reviewed-on: https://chromium-review.googlesource.com/234739
Reviewed-by: Alexandru Stan <amstan@chromium.org>