BACKPORT: FROMLIST: clk: rockchip: add clock ID for usbphy480m_src
NOTE: BACKPORT adjusts the clock ID.

---

There are 3 different parent clock from different usbphy,
all of them are fixed 480MHz, it is not able to auto select
by clock core to the 2nd and the 3rd parent.
For different use case for different board, we may need to
select different usbphy clock out as parent manually.

Add the clock ID for it so that we can use in dts.

BUG=chrome-os-partner:34667
TEST=Build and boot

Change-Id: I629280e2f2628051fcfd133840ac82fe3a69fb69
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/5293961/)
Reviewed-on: https://chromium-review.googlesource.com/229551