CHROMIUM: ASoC: rockchip: switch to using 227e interrupt
Using the interrupt line from the ts3a227e headset chip instead of the
individual hp/mic gpios will allow the system to see keys from watson
headsets.

MICBIAS needs to be enabled for the keys to work.  Force enable it so
it is high even when audio isn't playing (the play key is much more
useful that way).

BUG=none
TEST=plug unplug headset and watch events in evtest.

Change-Id: I2d3b5bac8208463dbbad6f542e874f59be709ef7
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/229874
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>