CHROMIUM: drm/rockchip: vop: disable vop irq when powered off
Just disabling vblank with "vblank_off" is not enough to prevent possibly
delayed or racing isrs (ie, executing on another CPU) from executing when
the VOP register access clocks is off.  To prevent this, we must call
disable_irq which ensures not only that there will be no future irqs, but
also it does not return until any currently running or pending irq has
completed.  Note that this also applies to the shared VOP IOMMU page fault
IRQ - disabling the shared IRQ also stops it from running when the master
VOP is off.  Note that we now start with the IRQ disabled, so we can
enable it initially when we first power on.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:33399
TEST=On veyron sanity check UI.
     powerd_dbus_suspend
TEST=w/ HDMI patches
     HDMI plug, then powerd_dbus_suspend, then wakeup
     HDMI unplug, then powerd_dbus_suspend, then wakeup
     HDMI plug, then powerd_dbus_suspend, then unplug, then wakeup
     HDMI unplug, then powerd_dbus_suspend, then plug, then wakeup

Change-Id: Ib0ab59889c733de6bda5a60fc992fcc82075dd48
Reviewed-on: https://chromium-review.googlesource.com/229229
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>