CHROMIUM: drm/rockchip: vop: add note about standby in power_off
The 'standby' command takes effect at the end of the current frame.
So, when we immediately shut off the scanout clock, it probably has not
yet taken effect.  Thus, the scanout was probably still active when we
killed its clock.

This shouldn't matter though, since in resume we restore iommu and then
clear standby.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=none

Change-Id: I59d0f6c5d1f90ace56333a1938e93558a999c151
Reviewed-on: https://chromium-review.googlesource.com/229228
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Reviewed-by: 智情 姚 <mark.yao@rock-chips.com>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>