HACK: clk: rockchip: Force rk3288 PWM clock (and its parents) on
Quick summary:

Most rk3288 boards (especially those with rk808 that followed the lead
from rk3288-evb-rk808) have a PWM regulator on them for vdd_logic.
This is the main voltage for all kinds of misc stuff including the
memory controller.

On these boards it is critically important to make sure that the PWM
never ever glitches and never loses its clock.  Any glitch could crash
the system.

Right now there are no users of the PWM regulator and also Linux
thinks that the PWM regulator is disabled.  Things happen to work
because firmware configured the PWM and Linux doesn't touch it.
...and the PWM's clock is marked as "ignore unused" (and also the PWM
code points to the wrong clock--see chrome-os-parnter:33455).

...but things _stop_ working if we turn off serial console.  Why?
Because:
1. Serial console shares a parent clock with the PWM (pclk_cpu)
2. If we have no serial console then nobody is holding pclk_cpu on
   at reboot time.  It gets disabled.

We need to fix a lot of the above problems, but until we get
everything right the cleanest "hack" seems like it is to just keep the
"rk_pwm" clock on always.

BUG=chrome-os-partner:33788, chrome-os-parnter:33455
TEST=Reboot test with console off.

Change-Id: I1377b04f9566c641694204c7d5f39126695a315e
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/230221
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>