CHROMIUM: Mali: wire up runtime PM support on rockchip
Wire up runtime pm support into rockchip platform portion of Mali
driver so that we can power on and off GPU power domain as needed.

Note that power operations are split in 2 parts: we disable/enable
clock synchronously, but disabling regulator and GPU power domain
is done asynchronously with 200 msec delay (using runtime autosuspend
framework).

BUG=chrome-os-partner:31652
TEST=boot to UI, inspect /sys/kernel/debug/pm_genpd/pm_genpd_summary and
observe that PM domain is going on an off depending on whether there is
a load on GPU.

Change-Id: Ib049c3273c2d74e3031fefa966ffdd2696fd19a5
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/230322
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>