CHROMIUM: ARM: dts: add i2s_clk_out to list of clocks used on veyron i2s
All veyron boards with the maxim codec will be driving the codec from
the i2s block, so tell the I2S block that it needs to claim that
output clock.  This also adds properties to model the relationship
between the output clock and the i2s clock itself, to ensure that the
clock muxing is set correctly.

BUG=chrome-os-partner:33886
TEST=audio works on pinky rev2

Change-Id: I7933e5d083cb40201943dcdc2fefb39e5fdb1c46
Signed-off-by: sonnyrao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/230349
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>