CHROMIUM: i915: prevent brightness from jumping up on enable
When a user dims the screen all the way, waits for a few seconds
so the backlight is disabled, and then raises the brightness, the
screen would flash on for a moment before the userspace daemons
changed it to the level it should be.

The i915 driver seems to be enforcing some brightness policy that
we don't like, though it would almost certainly break someone's
userspace if we floated this upstream.

BUG=chrome-os-partner:32459
TEST=manual

BUG=chrome-os-partner:47209
TEST=decrease brightness to level 0, after 1s panel will off, then
press any key,brightness from level 0 to level 1

Signed-off-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/230510
Reviewed-by: Stphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 394d0d0a1e82248f01ef5a35bc90ccc811f5d037)
Change-Id: I17af5582728e6d0dc6e053c856cf996ffd798f37
Signed-off-by: Gary Wang <gary.c.wang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/315940
Commit-Ready: Gary C Wang <gary.c.wang@intel.com>
Tested-by: Gary C Wang <gary.c.wang@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Cooper Chiou <cooper.chiou@intel.com>