CHROMIUM: power: bq2477x: Add monitor mode
Current bq2477x kernel driver can set charger characteristic
to the value specified in dts file and periodically control
that. However, we want to handle this operation to the EC.

This patch adds monitor mode to the driver that make it just
read the data and does not set anything.

Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>

BUG=chrome-os-partner:31931
TEST=/sys/class/power_supply/bq2477x-ac/online works in ryu P3

Change-Id: Ied6a7e53838ae600cff8327efc218c5153d1088c
Reviewed-on: https://chromium-review.googlesource.com/230624
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>