CHROMIUM: ryu: dts: Add pull up to KB_COL5 pin
We add signal ACOK_AP_L from charger to the AP at pin KB_COL5 in
ryu P3. This is an open drain signal so that it will require a
pull-up from the AP.

Also this pin does not connect in ryu P2 and earlier so it is safe
to add to common dts.

Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>

BUG=chrome-os-partner:32115
TEST=/sys/class/power_supply/bq2477x-ac/online works in ryu P3

Change-Id: I623e79bd620ed18170ff839e252a479a23b16ab8
Reviewed-on: https://chromium-review.googlesource.com/230625
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>