drm/locking: Allow NULL crtc in drm_modeset_legacy_acquire_ctx
I've missed checking this and so didn't notice that there's a NULL
check missing. Since depending upon calling context the crtc might not
even be there (disable-me-harder does happen around planes, especially
in cleanup code) we need to dodge the oops and look at the global
acquire ctx.

v2: Actually fix the oops for real and don't just move it two lines
down. That requires that we pass a drm_device pointer for the cases
where crtc could be NULL.

Reported-by: "Jasper St. Pierre" <jstpierre@mecheye.net>
Cc: "Jasper St. Pierre" <jstpierre@mecheye.net>
Cc: Rob Clark <robdclark@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>